+++ java/java-api/java-api/src/main/java/me/staek/stream/scenario/Stream_Ex2.java
@@ -1,4 +1,4 @@
-package me.staek.stream;
+package me.staek.stream.scenario;
 
 import java.util.stream.Stream;
